Clean up repository root layout
