--- .env
@@ -1,1 +1,0 @@
-PW_EXPERIMENTAL_SERVICE_WORKER_NETWORK_EVENTS=1
