+++ .gitignore
@@ -1,10 +1,15 @@
 .DS_Store
 
+/.codex-worktrees/
+/.env
+
 node_modules/
 
 builds/
 
 dictionaries/
+
+**/__pycache__/
 
 /test-results/
 /playwright-report/
