+++ docs/images/btn-add-duplicate-reading.png
