+++ docs/images/deinflection-example-simple.png
