+++ docs/images/firefox-marketplace.png
