+++ docs/images/ss-dictionaries.png
