+++ docs/images/ss-kanji-thumb.png
