+++ docs/images/ss-terms.png
