+++ playwright.config.js
@@ -23,6 +23,10 @@ import {defineConfig, devices} from '@playwright/test';
  */
 import 'dotenv/config';
 
+// Keep service-worker network events enabled even when contributors do not
+// have a local .env file in the repository root.
+process.env.PW_EXPERIMENTAL_SERVICE_WORKER_NETWORK_EVENTS ??= '1';
+
 const nodeMajorVersion = Number.parseInt(process.versions.node.split('.')[0], 10);
 if (!Number.isFinite(nodeMajorVersion) || nodeMajorVersion < 18 || nodeMajorVersion > 22) {
     throw new Error(
